Address rossy comments. Move command into its own files
