+++ package.json
@@ -15,8 +15,7 @@
     "tsx": "^4.7.1"
   },
   "dependencies": {
-    "@changesets/cli": "^2.26.2",
-    "zx": "^8.1.4"
+    "@changesets/cli": "^2.26.2"
   },
   "packageManager": "yarn@4.0.2",
   "private": true,
