+++ typescript/cli/package.json
@@ -19,7 +19,8 @@
     "yaml": "^2.4.1",
     "yargs": "^17.7.2",
     "zod": "^3.21.2",
-    "zod-validation-error": "^3.3.0"
+    "zod-validation-error": "^3.3.0",
+    "zx": "^8.1.4"
   },
   "devDependencies": {
     "@ethersproject/abi": "*",
